🚚 [RUMF-1097][rum] move RumInitConfiguration and HybridConfiguration
